+++ apps/client/app/routes/__auth/__admin/blog/$blogId.tsx
